# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de registro de asistencia",
        "description": "listado de URI's de la API alumnos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/cursos": {
            "get": {
                "tags": [
                    "Curso"
                ],
                "summary": "lista de todos los cursos",
                "operationId": "ce4902ce36ce58b9ed475eabcbf2dd16",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "curso": {
                                                        "type": "string",
                                                        "example": "Matematicas"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "hablamos de numeros"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Curso"
                ],
                "summary": "Registrar un nuevo curso",
                "operationId": "1ca42229715ee4cbdb9820b86e639989",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "curso": {
                                                "type": "string"
                                            },
                                            "descripcion": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "curso": "Matematica",
                                    "descripcion": "Hablamos de numeros"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "curso": {
                                            "type": "string",
                                            "example": "Matematica"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Hablamos de numeros"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "el campo curso es requerido."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cursos/{id}": {
            "put": {
                "tags": [
                    "Curso"
                ],
                "summary": "Actualizar la información de un curso",
                "operationId": "69cf34dcba8f354e6c95826a1c34004a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "curso": {
                                                "type": "string"
                                            },
                                            "descripcion": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "curso": "Matematica Editado",
                                    "descripcion": "Hablamos de numeros Editado"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "curso": {
                                            "type": "string",
                                            "example": "Matematica Editado"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Hablamos de numeros Editado"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "el campo curso es requerido."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Curso"
                ],
                "summary": "Eliminar la información de un alumno",
                "operationId": "6e380143cdb74c2815086d3935bff525",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}